backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
pydantic
python-dotenv
pypdf
python-docx
openpyxl
